requirements.txt
# API / Web interfaces
requests>=2.24.0

# Data analysis
numpy>=1.19.2
pandas>=1.1.3

# Testing
pytest>=6.1.1
vcrpy>=4.1.1

# Documentation
sphinx>=3.2.1
pydata-sphinx-theme>=0.4.1

# Secrets management
python-dotenv>=0.14.0
# TODO: Implement git-secret system for storage of secrets alongside repository

# Development
jupyter>=1.0.0